Cypress: Generate JWT tokens in a task
This is because this is done using `jsonwebtoken` which internally
import and uses Node's `util` which is no longer available out-of-the-box
in Cypress 15 (there was a polyfill for the test/browser environment before)
