+++ cypress.config.ts
@@ -1,4 +1,6 @@
 import { defineConfig } from 'cypress'
+import jwt from 'jsonwebtoken'
+
 import { resetStubs } from './integration_tests/mockApis/wiremock'
 
 import auth from './integration_tests/mockApis/auth'
@@ -9,6 +11,21 @@ import prisonApi from './integration_tests/mockApis/prisonApi'
 import offenderSearchApi from './integration_tests/mockApis/offenderSearchApi'
 import nonAssociationsApi from './integration_tests/mockApis/nonAssociationsApi'
 import frontendComponents from './integration_tests/mockApis/frontendComponents'
+
+function createToken({ roles }: { roles: string[] }) {
+  // authorities in the session are always prefixed by ROLE.
+  const authorities = roles.map(role => (role.startsWith('ROLE_') ? role : `ROLE_${role}`))
+  const payload = {
+    user_name: 'USER1',
+    scope: ['read', 'write'],
+    auth_source: 'NOMIS',
+    authorities,
+    jti: '83b50a10-cca6-41db-985f-e87efb303ddb',
+    client_id: 'clientid',
+  }
+
+  return jwt.sign(payload, 'secret', { expiresIn: '1h' })
+}
 
 export default defineConfig({
   chromeWebSecurity: false,
@@ -27,6 +44,7 @@ export default defineConfig({
     setupNodeEvents(on) {
       on('task', {
         resetStubs,
+        createToken,
         ...auth,
         ...manageUsersApi,
         ...tokenVerification,
